+++ src/index.html
@@ -36,6 +36,8 @@
     <footer class="container-fluid bg-secondary text-white text-center p-3">
         <p>&copy; 2023 Cooking Masters Sive and Raphael. All rights reserved.</p>
     </footer>
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     <script src="index.js" type="module"></script>
 </body>
