+++ src/style.css
@@ -62,6 +62,7 @@ footer {
   padding: 10px;
   left: 0;
   bottom: 0;
+  position: relative;
 }
 
 .modal-content {
